[S07P12D105-204] [BE] feat : 로그인 완료 시 반환 정보 추가

See merge request s07-webmobile1-sub2/S07P12D105!53

+++ BE/sellerB/src/main/java/backend/sellerB/entity/Manager.java
@@ -27,7 +27,7 @@ public class Manager implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     @Column(name = "manager_seq")
-    private int managerSeq;
+    private Long managerSeq;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "brand_seq")
     private Brand brandSeq;
